dom/base/test/test_meta_viewport_auto_size_by_invalid_width.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>invalid width enables autoSize</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=-1">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=-1</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function invalid_width() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      ok(info.autoSize, "invalid width should enable autoSize");
    });
  </script>
</body>
</html>
